[refactor]: Enum 분리
